requirements.txt
Pillow>=10.0
numpy>=1.26
matplotlib>=3.8
